Removed redundant test cases from the test spec
Rather than waste time and energy implementing test cases that we know are going to
pass, it is much more productive (and interesting) to implement test cases which may
require some ingenuity. In particular, implementing a test case to verify the
correct encoding of a binary file would be extremely useful.